# apps/api/plane/curve/contracts/openapi/curve-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Curve API
  version: 1.0.0
  license:
    name: GNU Affero General Public License v3.0 only
    identifier: AGPL-3.0-only
  description: >-
    Workspace-scoped Curve control-plane API. Every request is authorized against
    Plane workspace membership and Curve object policy before object retrieval.
servers:
  - url: /api/v1
security:
  - planeSession: []
paths:
...
  /workspaces/{workspace_slug}/curve/operations/{operation_id}:
    get:
      operationId: getCurveOperation
      summary: Read one operation
      parameters:
        - $ref: '#/components/parameters/WorkspaceSlug'
        - $ref: '#/components/parameters/OperationId'
      responses:
        '200':
          description: Current operation state
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '../schemas/operation-summary.schema.json'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    planeSession:
      type: apiKey
      in: cookie
      name: sessionid
      description: Plane authenticated session; mutating requests also require Plane's CSRF protection.
  parameters:
    WorkspaceSlug:
      name: workspace_slug
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 80
    OperationId:
      name: operation_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Problem:
      description: Curve Problem Details response
      content:
        application/problem+json:
          schema:
            $ref: '../schemas/common.schema.json#/$defs/ProblemDetails'